+++ com.unity.render-pipelines.universal/Runtime/UniversalRendererData.cs
@@ -93,6 +93,13 @@ namespace UnityEngine.Rendering.Universal
             [Reload("Shaders/Utils/FallbackLoading.shader")]
             public Shader fallbackLoadingPS;
 
+            /// <summary>
+            /// Material Error shader.
+            /// </summary>
+            [Obsolete("Use fallbackErrorPS instead")]
+            [Reload("Shaders/Utils/MaterialError.shader")]
+            public Shader materialErrorPS;
+
             // Core blitter shaders, adapted from HDRP
             // TODO: move to core and share with HDRP
             [Reload("Shaders/Utils/CoreBlit.shader"), SerializeField]
@@ -144,6 +151,7 @@ namespace UnityEngine.Rendering.Universal
         [SerializeField] bool m_ClusteredRendering = false;
         const TileSize k_DefaultTileSize = TileSize._32;
         [SerializeField] TileSize m_TileSize = k_DefaultTileSize;
+        [SerializeField] IntermediateTextureMode m_IntermediateTextureMode = IntermediateTextureMode.Always;
 
         protected override ScriptableRenderer Create()
         {
@@ -274,6 +282,19 @@ namespace UnityEngine.Rendering.Universal
                 Assert.IsTrue(value.IsValid());
                 SetDirty();
                 m_TileSize = value;
+            }
+        }
+
+        /// <summary>
+        /// Controls when URP renders via an intermediate texture.
+        /// </summary>
+        public IntermediateTextureMode intermediateTextureMode
+        {
+            get => m_IntermediateTextureMode;
+            set
+            {
+                SetDirty();
+                m_IntermediateTextureMode = value;
             }
         }